--- src/main/java/_3team/indev3teambackend/users/UserRepository.java
@@ -1,6 +1,0 @@
-package _3team.indev3teambackend.users;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
