+++ src/main/java/_3team/indev3teambackend/users/entity/User.java
@@ -1,13 +1,12 @@
-package _3team.indev3teambackend.users;
+package _3team.indev3teambackend.users.entity;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import _3team.indev3teambackend.ChatMessage.entity.ChatMessagesEntity;
+import _3team.indev3teambackend.ChatRoom.entity.ChatRoomEntity;
+import jakarta.persistence.*;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -40,6 +39,32 @@ public class User {
 
     @Column(name = "updated_at", insertable = false, updatable = false)
     private LocalDateTime updatedAt;
+
+
+    // ----------------------------------
+    // 외래 키(FK) 연결 - 관계 설정
+    // ----------------------------------
+
+    // 1. User와 ChatRooms 간의 1:N 관계 (사용자가 생성한 채팅방)
+    // "user" 필드는 ChatRoom 엔티티에서 User를 참조하는 필드 이름입니다.
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ChatRoomEntity> chatRooms = new ArrayList<>();
+
+    // 2. User와 ChatMessages 간의 1:N 관계 (사용자가 보낸 메시지)
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ChatMessagesEntity> chatMessages = new ArrayList<>();
+    // 생성 및 업데이트 시각 자동 설정
+    @PrePersist
+    public void onPrePersist() {
+        this.createdAt = LocalDateTime.now();
+        this.updatedAt = LocalDateTime.now();
+    }
+
+    @PreUpdate
+    public void onPreUpdate() {
+        this.updatedAt = LocalDateTime.now();
+    }
+
 
     public User(String providerId, String providerIdEmail, String name, String profileImageUrl) {
         this.providerId = providerId;
